(drivers) block_dev: Fix offset type for partitions

+++ src/drivers/block_dev/block_dev.h
@@ -37,7 +37,7 @@ struct block_dev {
 	struct dev_module *dev_module;
 
 	/* partitions */
-	size_t start_offset;
+	uint64_t start_offset;
 	struct block_dev *parent_bdev;
 };
